Miss reverted "Root fractal: first version working."
This reverts commit 7dc08bc24f4bb57c77e9e19ce6cf310afbba563a.

+++ src/support/initialization.rs
@@ -69,7 +69,7 @@ where
     {
         let dpi_mode = if let Ok(factor) = std::env::var("IMGUI_EXAMPLE_FORCE_DPI_FACTOR") {
             // Allow forcing of HiDPI factor for debugging purposes
-            match factor.parse::<f32>() {
+            match factor.parse::<f64>() {
                 Ok(f) => HiDpiMode::Locked(f),
                 Err(e) => panic!("Invalid scaling factor: {}", e),
             }
